requirements.txt
Django==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-unfold==0.34.0
djoser==2.3.1
drf-nested-routers==0.94.1
drf-yasg==1.21.10
django-filter==25.1
django-cors-headers==4.7.0
django-environ==0.11.2
django-redis==5.4.0
django-templated-mail==1.1.1
celery==5.4.0
redis==5.2.1
dj-database-url==2.3.0
psycopg2-binary==2.9.10
pillow==11.1.0
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.9.0
tzdata==2025.1
